fix: move getSeparator to format
