+++ packages/util/src/format/getSeparator.ts
