+++ packages/util/src/number/index.ts
@@ -5,6 +5,5 @@
  * @summary Utility methods to convert to and from `number` values
  */
 
-export { getSeparator } from './getSeparator';
 export { numberToHex } from './toHex';
 export { numberToU8a } from './toU8a';
